Move Container to its own file
