+++ packages/components/src/internal/app/models.ts
@@ -5,9 +5,7 @@
 import { Record } from 'immutable';
 import { ActionURL, getServerContext } from '@labkey/api';
 
-import { User } from '../..'
-
-import { Container } from '../components/base/models/model';
+import { Container, User } from '../..'
 
 const user = new User(getServerContext().user);
