+++ packages/components/src/internal/components/domainproperties/actions.ts
@@ -17,7 +17,8 @@ import classNames from 'classnames';
 import { List, Map } from 'immutable';
 import { Ajax, Domain, Query, Security, Utils } from '@labkey/api';
 
-import { Container, QueryColumn, SchemaDetails } from '../base/models/model';
+import { Container } from '../../..'
+import { QueryColumn, SchemaDetails } from '../base/models/model';
 import { naturalSort } from '../../util/utils';
 import { processSchemas } from '../base/models/schemas';
 import { buildURL } from '../../url/ActionURL';
