+++ packages/components/src/internal/components/domainproperties/assay/AssayPropertiesInput.tsx
@@ -3,7 +3,7 @@ import { Col, FormControl, Row } from 'react-bootstrap';
 import { List } from 'immutable';
 import { ActionURL } from '@labkey/api';
 
-import { Container } from '../../base/models/model';
+import { Container } from '../../../..';
 import { LoadingSpinner } from '../../base/LoadingSpinner';
 import { RemoveEntityButton } from '../../buttons/RemoveEntityButton';
 import { AddEntityButton } from '../../buttons/AddEntityButton';
